Main refactor pre debug
