+++ rr_viz/msg/RRTaskWaypoint.msg
@@ -1,4 +1,3 @@
-Header header
-geometry_msgs/PoseStamped pose
+geometry_msgs/PoseStamped pose_stamped
 std_msgs/String tasksrv
 std_msgs/String tasksubtree
