+++ rr_viz/src/rr_viz/rr_viz.ui
@@ -30,39 +30,11 @@
         <string>Inspect View</string>
        </attribute>
       </widget>
-      <widget class="QWidget" name="MissionViewTab">
+      <widget class="MissionControlView" name="MissionViewTab">
        <attribute name="title">
         <string>Mission View</string>
        </attribute>
-       <layout class="QHBoxLayout" name="horizontalLayout_2">
-        <item>
-         <widget class="RVizMinimap" name="mmWidget" native="true">
-          <property name="sizePolicy">
-           <sizepolicy hsizetype="Expanding" vsizetype="Preferred">
-            <horstretch>2</horstretch>
-            <verstretch>0</verstretch>
-           </sizepolicy>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <layout class="QVBoxLayout" name="verticalLayout">
-          <item>
-           <widget class="SlamSupervisorWidget" name="slamSupervisor" native="true"/>
-          </item>
-          <item>
-           <widget class="PathMakerWidget" name="pmWidget_2" native="true">
-            <property name="sizePolicy">
-             <sizepolicy hsizetype="Preferred" vsizetype="Expanding">
-              <horstretch>1</horstretch>
-              <verstretch>0</verstretch>
-             </sizepolicy>
-            </property>
-           </widget>
-          </item>
-         </layout>
-        </item>
-       </layout>
+       <layout class="QHBoxLayout" name="horizontalLayout_2"/>
       </widget>
       <widget class="ViewDiagnostics" name="DiagnosticsViewTab">
        <attribute name="title">
@@ -85,15 +57,9 @@
    </slots>
   </customwidget>
   <customwidget>
-   <class>PathMakerWidget</class>
+   <class>MissionControlView</class>
    <extends>QWidget</extends>
-   <header>widgets/path_maker</header>
-   <container>1</container>
-  </customwidget>
-  <customwidget>
-   <class>RVizMinimap</class>
-   <extends>QWidget</extends>
-   <header>widgets/rviz_minimap</header>
+   <header>widgets/mission_control/mission_control</header>
    <container>1</container>
   </customwidget>
   <customwidget>
@@ -112,12 +78,6 @@
    <class>RRQTabWidget</class>
    <extends>QTabWidget</extends>
    <header>widgets/rr_tab_widget</header>
-   <container>1</container>
-  </customwidget>
-  <customwidget>
-   <class>SlamSupervisorWidget</class>
-   <extends>QWidget</extends>
-   <header>widgets/slam_supervisor</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
