+++ rr_viz/src/rr_viz/widgets/networking/diagnostics_tab.ui
