+++ rr_viz/src/rr_viz/widgets/networking/status_machine.py
