+++ rr_viz/src/rr_viz/widgets/networking/view_diagnostics.py
